feat(install): add cleanup support for kustomizations
- Add cleanup support for kustomizations in Down method

- Update Down method to handle kustomization cleanup

- Add kustomization cleanup to cleanup process

- Update test coverage for kustomization cleanup
