+++ pkg/stack/windsor_stack.go
@@ -154,17 +154,17 @@ func (s *WindsorStack) Down() error {
 			}
 		}
 
-		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️ Initializing Terraform in %s", component.Path), "terraform", "init", "-migrate-state", "-upgrade", "-force-copy")
+		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️  Initializing Terraform in %s", component.Path), "terraform", "init", "-migrate-state", "-upgrade", "-force-copy")
 		if err != nil {
 			return fmt.Errorf("error initializing Terraform in %s: %w", component.FullPath, err)
 		}
 
-		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️ Planning Terraform destruction in %s", component.Path), "terraform", "plan", "-destroy")
+		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️  Planning Terraform destruction in %s", component.Path), "terraform", "plan", "-destroy")
 		if err != nil {
 			return fmt.Errorf("error planning Terraform destruction in %s: %w", component.FullPath, err)
 		}
 
-		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️ Destroying Terraform resources in %s", component.Path), "terraform", "destroy", "-auto-approve")
+		_, err = s.shell.ExecProgress(fmt.Sprintf("🗑️  Destroying Terraform resources in %s", component.Path), "terraform", "destroy", "-auto-approve")
 		if err != nil {
 			return fmt.Errorf("error destroying Terraform resources in %s: %w", component.FullPath, err)
 		}
